PVS-Studio check

+++ RowDaBoat-x64barebones-d4e1c147f975/Userland/SampleCodeModule/include/stdio.h
@@ -14,7 +14,7 @@ void printText(char *string);
 
 void putChar(char character);
 
-uint8_t getChar();
+int getChar();
 
 void newline();
